refactor why trade and solve qa issues
